+++ sdk/communication/azure-communication-chat/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 ## 1.0.0-beta.9 (Unreleased)
+### Breaking Changes
+- Removed `com.azure.android.communication.chat.signaling.chatevents.ChatParticipant`
+- Removed `com.azure.android.communication.chat.signaling.chatevents.ChatThreadProperties`
+- Changed `com.azure.android.communication.chat.signaling.chatevents.*` to `com.azure.android.communication.chat.models.signaling.*`
 
 ## 1.0.0-beta.8 (2021-03-29)
 ### Breaking Changes